backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
cryptography>=42.0.0
httpx>=0.27.0
websockets>=12.0
torch>=2.3.0
transformers>=4.41.0
sentence-transformers>=3.0.0
faiss-cpu>=1.13.0
shap>=0.45.0
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.14.0
numpy>=1.26.0
torchaudio>=2.3.0
ffmpeg-python>=0.2.0
faster-whisper>=1.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
aiofiles>=23.2.0
setuptools>=70.0.0
email-validator>=2.0.0
greenlet>=3.0.0
bcrypt==4.0.1
google-generativeai>=0.5.0
fastapi-mail>=1.4.0
aiosmtplib>=3.0.0
jinja2>=3.1.0
groq==0.18.0